style: move pair hash fn into strong types file
