+++ src/utils/mmbot_strong_types.hpp
@@ -79,4 +79,22 @@ namespace antara
         std::pair<asset, asset> to_std_pair();
         bool operator==(const pair &rhs) const;
     };
+
+}
+
+namespace std {
+
+template <>
+struct hash<antara::pair> {
+  std::size_t operator()(const antara::pair& p) const {
+    using std::size_t;
+    using std::hash;
+
+    std::size_t h1 = std::hash<std::string>{}(p.base.symbol.value());
+    std::size_t h2 = std::hash<std::string>{}(p.quote.symbol.value());
+
+    return h1 ^ (h2 << 1);
+  }
+};
+
 }
